+++ extensions/bland/webhooks/index.ts
@@ -1,3 +1,4 @@
 import { callCompleted } from './CallCompleted/callCompleted'
+import { completedBlandText } from './CompletedBlandText/completedBlandText'
 
-export const webhooks = [callCompleted]
+export const webhooks = [callCompleted, completedBlandText]