+++ extensions/metriport/actions/index.ts
@@ -9,6 +9,7 @@ import { startNetworkQuery } from './network/startNetworkQuery'
 import { startConsolidatedQuery } from './consolidated/startConsolidatedQuery'
 import { getConsolidatedQueryStatus } from './consolidated/getConsolidatedQueryStatus'
 import { getWebhookBundle } from './webhookBundle/getWebhookBundle'
+import { removePatientFromCohort } from './cohort/removePatientFromCohort'
 
 export const actions = {
   getPatient,
@@ -22,4 +23,5 @@ export const actions = {
   startConsolidatedQuery,
   getConsolidatedQueryStatus,
   getWebhookBundle,
+  removePatientFromCohort,
 }